refactor based on PR #753 (round 1)
